+++ src/app/modules/lux-layout/lux-app-footer/lux-app-footer-button-info.ts
@@ -3,6 +3,7 @@ import { LuxActionColorType } from "../../lux-action/lux-action-model/lux-action
 export interface ILuxAppFooterButtonInfo {
   label: string;
   cmd: string;
+  prio: number;
   color?: LuxActionColorType;
   disabled?: boolean;
   hidden?: boolean;
@@ -16,6 +17,7 @@ export interface ILuxAppFooterButtonInfo {
 export class LuxAppFooterButtonInfo implements ILuxAppFooterButtonInfo {
   label: string;
   color: LuxActionColorType;
+  prio: number;
   disabled: boolean;
   cmd: string;
   hidden: boolean;
@@ -28,6 +30,7 @@ export class LuxAppFooterButtonInfo implements ILuxAppFooterButtonInfo {
   constructor(
     label: string,
     cmd: string,
+    prio?: number,
     color?: LuxActionColorType,
     disabled?: boolean,
     hidden?: boolean,
@@ -39,13 +42,13 @@ export class LuxAppFooterButtonInfo implements ILuxAppFooterButtonInfo {
   ) {
     this.label = label;
     this.color = color;
+    this.prio = !prio ? 0 : prio;
     this.disabled = disabled;
     this.cmd = cmd;
     this.hidden = hidden === true ? true : false;
     this.raised = raised === undefined || raised === null || raised === true ? true : false;
     this.iconName = iconName;
-    this.alwaysVisible =
-      alwaysVisible === undefined || alwaysVisible === null || alwaysVisible === false ? false : true;
+    this.alwaysVisible = alwaysVisible === undefined || alwaysVisible === null || alwaysVisible === false ? false : true;
     this.tooltip = tooltip ? tooltip : '';
     this.onClick = onClick ? onClick : (that: ILuxAppFooterButtonInfo) => {};
   }
@@ -57,20 +60,11 @@ export class LuxAppFooterButtonInfo implements ILuxAppFooterButtonInfo {
    * @param data
    * @returns eine Button Info
    */
-  static generateInfo(data: ILuxAppFooterButtonInfo): LuxAppFooterButtonInfo {
-    const info: LuxAppFooterButtonInfo = new LuxAppFooterButtonInfo(
-      data.label,
-      data.cmd,
-      data.color,
-      data.disabled,
-      data.hidden,
-      data.raised,
-      data.iconName,
-      data.alwaysVisible,
-      data.tooltip,
-      data.onClick
-    );
+  static generateInfo(data: Partial<ILuxAppFooterButtonInfo>): LuxAppFooterButtonInfo {
+    const info = new LuxAppFooterButtonInfo(data.label, data.cmd);
+    Object.assign(info, data);
 
     return info;
   }
+
 }